Должно работать обновление

+++ sources/PLoader/src/Hardware/HAL/HAL_EEPROM.cpp
@@ -12,7 +12,7 @@
                             FLASH_FLAG_PGSERR);  /* programming sequence error flag    */
 
 
-void HAL_EEPROM::WriteBuffer(uint address, uint8 *data, int size)
+void HAL_EEPROM::WriteBuffer(int address, uint8 *data, int size)
 {
     CLEAR_FLASH_FLAGS;
 
@@ -20,7 +20,7 @@ void HAL_EEPROM::WriteBuffer(uint address, uint8 *data, int size)
 
     for(int i = 0; i < size; i++)
     {
-        HAL_FLASH_Program(TYPEPROGRAM_BYTE, address, data[i]);
+        HAL_FLASH_Program(TYPEPROGRAM_BYTE, static_cast<uint>(address), data[i]);
         ++address;
     }
